vhost: fix IPv4 csum calculation
Currently the IPv4 header checksum is calculated including
the its current value, which can be a valid checksum or just
garbage. In any case, if the original value is not zero, then
the result is always wrong.

The IPv4 checksum is defined in RFC791, page 14 says:
  Header Checksum:  16 bits

  The checksum algorithm is:
  The checksum field is the 16 bit one's complement of the one's
  complement sum of all 16 bit words in the header.  For purposes of
  computing the checksum, the value of the checksum field is zero.

Fix that to always force the field to be zero.

Signed-off-by: Flavio Leitner <fbl@sysclose.org>